+++ lib/src/components/resizable.dart
@@ -1,7 +1,7 @@
 import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:lucide_icons_flutter/lucide_icons.dart';
-import 'package:shadcn_ui/src/components/divider.dart';
+import 'package:shadcn_ui/src/components/separator.dart';
 import 'package:shadcn_ui/src/theme/components/decorator.dart';
 import 'package:shadcn_ui/src/theme/theme.dart';
 import 'package:shadcn_ui/src/utils/border.dart';
@@ -529,7 +529,7 @@ class ShadResizablePanelGroupState extends State<ShadResizablePanelGroup> {
         }
 
         final divider = switch (widget.axis) {
-          Axis.horizontal => ShadDivider.vertical(
+          Axis.horizontal => ShadSeparator.vertical(
               margin:
                   EdgeInsets.symmetric(horizontal: effectiveDividerSize / 2),
               thickness: effectiveDividerThickness,
@@ -538,7 +538,7 @@ class ShadResizablePanelGroupState extends State<ShadResizablePanelGroup> {
           Axis.vertical => SizedBox(
               // double.infinity doesn't work, just providing a big number
               width: 50000,
-              child: ShadDivider.horizontal(
+              child: ShadSeparator.horizontal(
                 margin:
                     EdgeInsets.symmetric(vertical: effectiveDividerSize / 2),
                 thickness: effectiveDividerThickness,