+++ lib/src/theme/data.dart
@@ -14,7 +14,6 @@ import 'package:shadcn_ui/src/theme/components/context_menu.dart';
 import 'package:shadcn_ui/src/theme/components/date_picker.dart';
 import 'package:shadcn_ui/src/theme/components/decorator.dart';
 import 'package:shadcn_ui/src/theme/components/dialog.dart';
-import 'package:shadcn_ui/src/theme/components/divider.dart';
 import 'package:shadcn_ui/src/theme/components/input.dart';
 import 'package:shadcn_ui/src/theme/components/input_otp.dart';
 import 'package:shadcn_ui/src/theme/components/menubar.dart';
@@ -24,6 +23,7 @@ import 'package:shadcn_ui/src/theme/components/progress.dart';
 import 'package:shadcn_ui/src/theme/components/radio.dart';
 import 'package:shadcn_ui/src/theme/components/resizable.dart';
 import 'package:shadcn_ui/src/theme/components/select.dart';
+import 'package:shadcn_ui/src/theme/components/separator.dart';
 import 'package:shadcn_ui/src/theme/components/sheet.dart';
 import 'package:shadcn_ui/src/theme/components/slider.dart';
 import 'package:shadcn_ui/src/theme/components/switch.dart';
@@ -93,7 +93,7 @@ class ShadThemeData extends ShadBaseTheme {
     ShadTimePickerTheme? timePickerTheme,
     ShadInputOTPTheme? inputOTPTheme,
     ShadMenubarTheme? menubarTheme,
-    ShadDividerTheme? dividerTheme,
+    ShadSeparatorTheme? separatorTheme,
   }) {
     final effectiveRadius =
         radius ?? const BorderRadius.all(Radius.circular(6));
@@ -237,7 +237,8 @@ class ShadThemeData extends ShadBaseTheme {
           effectiveVariant.timePickerTheme().mergeWith(timePickerTheme),
       inputOTPTheme: effectiveVariant.inputOTPTheme().mergeWith(inputOTPTheme),
       menubarTheme: effectiveVariant.menubarTheme().mergeWith(menubarTheme),
-      dividerTheme: effectiveVariant.dividerTheme().mergeWith(dividerTheme),
+      separatorTheme:
+          effectiveVariant.separatorTheme().mergeWith(separatorTheme),
     );
   }
 
@@ -292,7 +293,7 @@ class ShadThemeData extends ShadBaseTheme {
     required super.timePickerTheme,
     required super.inputOTPTheme,
     required super.menubarTheme,
-    required super.dividerTheme,
+    required super.separatorTheme,
   });
 
   static ShadThemeData lerp(ShadThemeData a, ShadThemeData b, double t) {
@@ -396,7 +397,8 @@ class ShadThemeData extends ShadBaseTheme {
       inputOTPTheme:
           ShadInputOTPTheme.lerp(a.inputOTPTheme, b.inputOTPTheme, t),
       menubarTheme: ShadMenubarTheme.lerp(a.menubarTheme, b.menubarTheme, t),
-      dividerTheme: ShadDividerTheme.lerp(a.dividerTheme, b.dividerTheme, t),
+      separatorTheme:
+          ShadSeparatorTheme.lerp(a.separatorTheme, b.separatorTheme, t),
     );
   }
 
@@ -455,7 +457,7 @@ class ShadThemeData extends ShadBaseTheme {
         other.timePickerTheme == timePickerTheme &&
         other.inputOTPTheme == inputOTPTheme &&
         other.menubarTheme == menubarTheme &&
-        other.dividerTheme == dividerTheme;
+        other.separatorTheme == separatorTheme;
   }
 
   @override
@@ -510,7 +512,7 @@ class ShadThemeData extends ShadBaseTheme {
         timePickerTheme.hashCode ^
         inputOTPTheme.hashCode ^
         menubarTheme.hashCode ^
-        dividerTheme.hashCode;
+        separatorTheme.hashCode;
   }
 
   ShadThemeData copyWith({
@@ -564,7 +566,7 @@ class ShadThemeData extends ShadBaseTheme {
     ShadTimePickerTheme? timePickerTheme,
     ShadInputOTPTheme? inputOTPTheme,
     ShadMenubarTheme? menubarTheme,
-    ShadDividerTheme? dividerTheme,
+    ShadSeparatorTheme? separatorTheme,
   }) {
     return ShadThemeData(
       colorScheme: colorScheme ?? this.colorScheme,
@@ -622,7 +624,7 @@ class ShadThemeData extends ShadBaseTheme {
       timePickerTheme: timePickerTheme ?? this.timePickerTheme,
       inputOTPTheme: inputOTPTheme ?? this.inputOTPTheme,
       menubarTheme: menubarTheme ?? this.menubarTheme,
-      dividerTheme: dividerTheme ?? this.dividerTheme,
+      separatorTheme: separatorTheme ?? this.separatorTheme,
     );
   }
 }