+++ playground/lib/pages/divider.dart
@@ -2,13 +2,13 @@ import 'package:flutter/material.dart';
 
 import 'package:shadcn_ui/shadcn_ui.dart';
 
-class DividerPage extends StatelessWidget {
-  const DividerPage({
+class SeparatorPage extends StatelessWidget {
+  const SeparatorPage({
     super.key,
     required this.style,
   });
 
-  final ShadDividerVariant style;
+  final ShadSeparatorVariant style;
 
   @override
   Widget build(BuildContext context) {
@@ -17,12 +17,12 @@ class DividerPage extends StatelessWidget {
         child: Padding(
           padding: const EdgeInsets.symmetric(horizontal: 40),
           child: switch (style) {
-            ShadDividerVariant.horizontal => const ShadDivider.horizontal(
+            ShadSeparatorVariant.horizontal => const ShadSeparator.horizontal(
                 thickness: 4,
                 margin: EdgeInsets.symmetric(horizontal: 20),
                 radius: BorderRadius.all(Radius.circular(4)),
               ),
-            ShadDividerVariant.vertical => const ShadDivider.vertical(
+            ShadSeparatorVariant.vertical => const ShadSeparator.vertical(
                 thickness: 4,
                 margin: EdgeInsets.symmetric(vertical: 20),
                 radius: BorderRadius.all(Radius.circular(4)),