+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.22.0
+
+- **FEAT**:: Add `ShadMenubar` component.
+- **REFACTOR**: Deprecate `ShadDivider` and `ShadDividerTheme`, use `ShadSeparator` and `ShadSeparatorTheme` instead.
+
 ## 0.21.0
 
 - **FEAT**: Add `ShadDivider` component (thanks to @Luckey-Elijah)